+++ src/components.rs
@@ -8,8 +8,8 @@ use tokio::sync::mpsc::UnboundedSender;
 use crate::{action::Action, config::Config, tui::Event};
 
 pub mod api_health;
-pub mod command_log;
 pub mod health;
+pub mod log_pane;
 pub mod lottery;
 pub mod network;
 pub mod peers;